+++ blocklets/ai-studio/src/pages/project/publish/publish-button.tsx
@@ -1,5 +1,6 @@
 import LoadingButton from '@app/components/loading/loading-button';
 import { useCurrentProject } from '@app/contexts/project';
+import { useIsAdmin } from '@app/contexts/session';
 import { AI_STUDIO_COMPONENT_DID } from '@app/libs/constants';
 import { getProjectIconUrl } from '@app/libs/project';
 import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
@@ -20,6 +21,9 @@ export default function PublishButton({ ...props }: LoadingButtonProps) {
   } = useProjectState(projectId, projectRef);
 
   const [showCreateResource, setShowCreateResource] = useState(false);
+  const isAdmin = useIsAdmin();
+
+  if (window.blocklet.DISABLE_AI_STUDIO_PUBLISH === 'true' && !isAdmin) return null;
 
   return (
     <>
@@ -49,11 +53,14 @@ function PublishDialog({ onClose }: { onClose: () => void }) {
     state: { project },
   } = useProjectState(projectId, projectRef);
 
+  const [opened, setOpened] = useState(false);
+
   if (!project) return null;
 
   return (
     <Suspense>
       <BlockletStudio
+        style={{ opacity: opened ? 1 : 0 }}
         mode="dialog"
         tenantScope={projectId}
         title={project.name || ''}
@@ -70,7 +77,7 @@ function PublishDialog({ onClose }: { onClose: () => void }) {
         onConnected={() => {}}
         onUploaded={() => {}}
         onReleased={() => {}}
-        onOpened={() => {}}
+        onOpened={() => setOpened(true)}
         // 默认选中的资源
         resources={{
           [AI_STUDIO_COMPONENT_DID]: [],